Peronnel Add and Edit not working well at the moment.             Add personnel related functionalities: list, create, update and fix uniques in personnel model
The commit includes additions of functionalities related to teacher. We added ways to list, create and update teachers. We also updated the day190 reporting and added ajax endpoint to fetch teacher's email. Moreover, the commit fixes unique constraints on personnel model in django and added support for formset javascript.

Other changes in this commit include modification in css and style for the better user interface. The views were updated along with url paths for the new functionalities. These changes aim to improve user experience and provide more efficient management of teachers data.