Add phone number field, update report generation and clean code
This commit includes multiple changes including 1) Addition of a phone number field for personnel which is optional and can be up to 25 characters long, handled in migrations. 2) Changes to the employee report feature. Reports are now ordered by 'last_name' and includes updates to the method handling save/complete actions for the report. 3) A command is added that calculates the log for student occurrences in countries and 4) Years_at_this_school field in the personnel model now defaults to 1. 5) Countries are now ordered based on the log of student occurrence and name. It also involves some code cleaning tasks across multiple files.