Update models and forms to allow changing active school year
The School model has been updated to include a reference to the active school year. A form and corresponding view have been added to allow users to change the active school year. Relevant model migrations are included. Changes to the context processor and templates have also been made to display the selected school year in the navbar.