Refactor form processing and update settings
This commit streamlines the process of saving forms in services/views.py. Now forms are checked for validity collectively and saved together if valid. It also moves the generation of a school year form from users/views.py to its own context processor in ISEIapp/context_processors.py, resulting in an update to ISEIapp/settings.py to include this new processor.