Add part-time student status and counts to reporting system
This update introduces a 'part-time' status for students in the reporting system. Part-time students are now included in student reports and counted separately under their own grade cound. A PartTimeGradeCount model was created to hold these counts. Related changes were made in the report views and templates. This commit also contains a minor change to an error message and the footer address.